requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
yt-dlp>=2025.1.1
python-dotenv>=1.0.1
requests>=2.32.0
pydantic>=2.8.0
openai>=1.40.0
